requirements.txt
fastapi==0.73.0
Pillow==9.0.0
pymongo==4.0.1
torch==1.9.0
torchvision==0.10.0
uvicorn==0.17.1
python-multipart==0.0.5
pydantic==1.9.0
